refactor: unsafe, CStr and libc moved out
