+++ deltachat-ffi/Cargo.toml
@@ -20,6 +20,7 @@ deltachat-provider-database = "0.2.1"
 libc = "0.2"
 human-panic = "1.0.1"
 num-traits = "0.2.6"
+failure = "0.1.6"
 
 [features]
 default = ["vendored", "nightly", "ringbuf"]
